Performing a simplifying transformation on the PGP gVCF data.
Transform the data such that all calls relevant to a particular allele are now in the same record.

This has a few benefits:
(1) the records are now bi-allelic, so simpler
(2) since the merge criteria is just a string match on (chrom, pos, ref, alt, end) the merge logic is quite straigtforward
(3) the schema itself is simpler because several of the repeating fields are now scalars (BigQuery treats all repeating fields as indpendent, but the ones we are converting to scalar here are *dependent*, they had a value per alternate allele if not null)

The data when imported directly from Variant Store to BigQuery closely mirrors the VCF file format.  The transformation performed here creates data deviates from the VCF format in one important way.  For any genotype value of 2 in a record, one cannot retrieve the value of that alternate allele from that record.  That alternate allele will be found in a different record.